+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.7] - 20.10.2023
+### Added
+- Added repository layer with short url decoding and encoding methods.
+### Modified
+- Now service returns short url instead of index.
+
 ## [0.1.6] - 19.10.2023
 ### Added
 - Added URL validity check.